переносит функцию отрисовки карточек в метод render
